backend/consumers/shipment_consumer/requirements.txt
kafka-python
sqlalchemy
psycopg2-binary
pydantic
python-dotenv